+++ lib/websocket.js
@@ -6,6 +6,7 @@ module.exports = {
     'request'      : require('./WebSocketRequest'),
     'connection'   : require('./WebSocketConnection'),
     'w3cwebsocket' : require('./W3CWebSocket'),
+    'connection2w3cwebsocket': require('./W3CWebSocketWrapper').connToW3C,
     'deprecation'  : require('./Deprecation'),
     'version'      : require('./version')
 };
